feat(idea): update ignore list
