+++ .idea/GitCommitMessageStorage.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="GitCommitMessageStorage">
+    <option name="messageStorage">
+      <MessageStorage />
+    </option>
+  </component>
+</project>
